+++ libft/src/ft_strlen_null.c
@@ -1,20 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_strlen_null.c                                   :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: svan-hoo <svan-hoo@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/03/23 20:26:24 by svan-hoo          #+#    #+#             */
-/*   Updated: 2024/11/22 20:47:49 by svan-hoo         ###   ########.fr       */
+/*                                                        ::::::::            */
+/*   ft_strlen_null.c                                   :+:    :+:            */
+/*                                                     +:+                    */
+/*   By: svan-hoo <svan-hoo@student.42.fr>            +#+                     */
+/*                                                   +#+                      */
+/*   Created: 2024/03/23 20:26:24 by svan-hoo      #+#    #+#                 */
+/*   Updated: 2024/12/12 18:36:17 by jdobos        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-int	ft_strlen_null(const char *str)
+size_t	ft_strlen_null(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (str)
